+++ api/v1alpha1/ai_service_backend.go
@@ -1,0 +1,72 @@
+// Copyright Envoy AI Gateway Authors
+// SPDX-License-Identifier: Apache-2.0
+// The full text of the Apache license is available in the LICENSE file at
+// the root of the repo.
+
+package v1alpha1
+
+import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	gwapiv1 "sigs.k8s.io/gateway-api/apis/v1"
+)
+
+// AIServiceBackend is a resource that represents a single backend for AIGatewayRoute.
+// A backend is a service that handles traffic with a concrete API specification.
+//
+// A AIServiceBackend is "attached" to a Backend which is either a k8s Service or a Backend resource of the Envoy Gateway.
+//
+// When a backend with an attached AIServiceBackend is used as a routing target in the AIGatewayRoute (more precisely, the
+// HTTPRouteSpec defined in the AIGatewayRoute), the ai-gateway will generate the necessary configuration to do
+// the backend specific logic in the final HTTPRoute.
+//
+// +kubebuilder:object:root=true
+// +kubebuilder:subresource:status
+// +kubebuilder:printcolumn:name="Status",type=string,JSONPath=`.status.conditions[-1:].type`
+type AIServiceBackend struct {
+	metav1.TypeMeta   `json:",inline"`
+	metav1.ObjectMeta `json:"metadata,omitempty"`
+	// Spec defines the details of AIServiceBackend.
+	Spec AIServiceBackendSpec `json:"spec,omitempty"`
+	// Status defines the status details of the AIServiceBackend.
+	Status AIServiceBackendStatus `json:"status,omitempty"`
+}
+
+// AIServiceBackendList contains a list of AIServiceBackends.
+//
+// +kubebuilder:object:root=true
+type AIServiceBackendList struct {
+	metav1.TypeMeta `json:",inline"`
+	metav1.ListMeta `json:"metadata,omitempty"`
+	Items           []AIServiceBackend `json:"items"`
+}
+
+// AIServiceBackendSpec details the AIServiceBackend configuration.
+type AIServiceBackendSpec struct {
+	// APISchema specifies the API schema of the output format of requests from
+	// Envoy that this AIServiceBackend can accept as incoming requests.
+	// Based on this schema, the ai-gateway will perform the necessary transformation for
+	// the pair of AIGatewayRouteSpec.APISchema and AIServiceBackendSpec.APISchema.
+	//
+	// This is required to be set.
+	//
+	// +kubebuilder:validation:Required
+	APISchema VersionedAPISchema `json:"schema"`
+	// BackendRef is the reference to the Backend resource that this AIServiceBackend corresponds to.
+	//
+	// A backend must be a Backend resource of Envoy Gateway. Note that k8s Service will be supported
+	// as a backend in the future.
+	//
+	// This is required to be set.
+	//
+	// +kubebuilder:validation:Required
+	BackendRef gwapiv1.BackendObjectReference `json:"backendRef"`
+
+	// BackendSecurityPolicyRef is the name of the BackendSecurityPolicy resources this backend
+	// is being attached to.
+	//
+	// +optional
+	BackendSecurityPolicyRef *gwapiv1.LocalObjectReference `json:"backendSecurityPolicyRef,omitempty"`
+
+	// TODO: maybe add backend-level LLMRequestCost configuration that overrides the AIGatewayRoute-level LLMRequestCost.
+	// 	That may be useful for the backend that has a different cost calculation logic.
+}
